add the option to seek named breaks, not just taken breaks; refactor headers to make this practical
