+++ src/passes/MergeBlocks.cpp
@@ -63,8 +63,9 @@
 
 #include <wasm.h>
 #include <pass.h>
-#include <ast_utils.h>
 #include <wasm-builder.h>
+#include <ast_utils.h>
+#include <ast/effects.h>
 
 namespace wasm {
