+++ src/passes/SimplifyLocals.cpp
@@ -45,8 +45,8 @@
 #include <wasm-builder.h>
 #include <wasm-traversal.h>
 #include <pass.h>
-#include <ast_utils.h>
 #include <ast/count.h>
+#include <ast/effects.h>
 #include <ast/manipulation.h>
 
 namespace wasm {
