+++ test/passes/vacuum.txt
@@ -5,6 +5,7 @@
  (type $3 (func (result i32)))
  (type $4 (func (param i32 f64 i32 i32)))
  (type $FUNCSIG$i (func (result i32)))
+ (type $6 (func (result f64)))
  (import "env" "int" (func $int (result i32)))
  (global $Int i32 (i32.const 0))
  (memory $0 256 256)
@@ -217,6 +218,19 @@
     (f64.const -nan:0xfffffffffa361)
     (loop $label$1
      (br $label$1)
+    )
+   )
+  )
+ )
+ (func $leave-block-even-if-br-not-taken (type $6) (result f64)
+  (block $label$0
+   (f64.store align=1
+    (i32.const 879179022)
+    (br_if $label$0
+     (loop $label$9
+      (br $label$9)
+     )
+     (i32.const 677803374)
     )
    )
   )
